调整侧边栏内添加网址代码

+++ js/event/website.event.js
@@ -3,14 +3,16 @@ import {
     renderCommonUse,
     addCommon,
     changeCommon,
-    deleteCommon
+    deleteCommon,
+    addSideBarWebsite,
+    createWebsite
 } from "../module/website.func.js"
 
 export const handleWebsite = (data = {}, state = '', isSetting = false) => {
     return new Promise((resolve, reject) => {
         //修改常用网址计数 data数据 state状态 isSetting是否设置
         //计数
-        if (state == 'count') {
+        if (state == 'count' && data.source == 'commons') {
             changeCommonCount(data)
                 .then(res => {
                     resolve(res);
@@ -19,7 +21,7 @@ export const handleWebsite = (data = {}, state = '', isSetting = false) => {
                 });
         }
         //添加
-        if (state == 'add') {
+        if (state == 'add' && data.source == 'commons') {
             addCommon(data)
                 .then(res => {
                     resolve(res);
@@ -28,7 +30,7 @@ export const handleWebsite = (data = {}, state = '', isSetting = false) => {
                 });
         }
         //修改
-        if (state == 'change') {
+        if (state == 'change' && data.source == 'commons') {
             changeCommon(data)
                 .then(res => {
                     resolve(res);
@@ -37,7 +39,7 @@ export const handleWebsite = (data = {}, state = '', isSetting = false) => {
                 })
         }
         //删除
-        if (state == 'delete') {
+        if (state == 'delete' && data.source == 'commons') {
             deleteCommon(data)
                 .then(res => {
                     resolve(res);
@@ -45,7 +47,20 @@ export const handleWebsite = (data = {}, state = '', isSetting = false) => {
                     reject(err);
                 })
         }
+        if (state == 'add' && data.source == 'addCapsule') {
+            addSideBarWebsite(data)
+                .then(res => {
+                    resolve(res);
+                }).catch(err => {
+                    reject(err);
+                })
+        }
         //渲染常用网址
-        renderCommonUse(isSetting);
+        if (data.source == 'commons') {
+            renderCommonUse(isSetting);
+        }
+        if (data.source == 'addCapsule') {
+            scrollContent.innerHTML = createWebsite();
+        }
     })
 }
